+++ jkube-kit/resource/service/src/test/java/org/eclipse/jkube/kit/config/service/openshift/OpenshiftBuildServiceTest.java
@@ -41,6 +41,8 @@ import org.eclipse.jkube.kit.common.JKubeProject;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.config.image.build.OpenShiftBuildStrategy;
 import org.eclipse.jkube.kit.config.resource.BuildRecreateMode;
+import org.eclipse.jkube.kit.config.resource.OpenshiftBuildConfig;
+import org.eclipse.jkube.kit.config.resource.ResourceConfig;
 import org.eclipse.jkube.kit.config.service.BuildService;
 import org.eclipse.jkube.kit.config.service.JKubeServiceException;
 import mockit.Expectations;
@@ -56,14 +58,15 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-
 
 public class OpenshiftBuildServiceTest {
 
@@ -449,6 +452,40 @@ public class OpenshiftBuildServiceTest {
                 lines = IOUtils.readLines(reader);
             }
             assertTrue(lines.contains("FOO=BAR"));
+        });
+    }
+
+    @Test
+    public void testBuildConfigResourceConfig() throws Exception {
+        retryInMockServer(() -> {
+            Map<String, String> limitsMap = new HashMap<>();
+            limitsMap.put("cpu", "100m");
+            limitsMap.put("memory", "256Mi");
+
+            BuildService.BuildServiceConfig config = defaultConfig
+                    .resourceConfig(new ResourceConfig.Builder()
+                            .withOpenshiftBuildConfig(new OpenshiftBuildConfig.Builder().limits(limitsMap).build()).build()).build();
+            OpenShiftMockServer mockServer = new OpenShiftMockServer();
+
+            OpenShiftClient client = mockServer.createOpenShiftClient();
+            final OpenshiftBuildService service = new OpenshiftBuildService(client, logger, dockerServiceHub, config);
+
+            ImageConfiguration imageWithEnv = new ImageConfiguration.Builder(image)
+                    .buildConfig(new JKubeBuildConfiguration.Builder(image.getBuildConfiguration())
+                            .env(Collections.singletonMap("FOO", "BAR"))
+                            .build()
+                    ).build();
+
+            KubernetesListBuilder builder = new KubernetesListBuilder();
+            service.createBuildArchive(imageWithEnv);
+            service.updateOrCreateBuildConfig(config, client, builder, imageWithEnv, null);
+            BuildConfig buildConfig = (BuildConfig) builder.buildFirstItem();
+            assertNotNull(buildConfig);
+            assertNotNull(buildConfig.getSpec().getResources());
+            assertEquals("256", buildConfig.getSpec().getResources().getLimits().get("memory").getAmount());
+            assertEquals("Mi", buildConfig.getSpec().getResources().getLimits().get("memory").getFormat());
+            assertEquals("100", buildConfig.getSpec().getResources().getLimits().get("cpu").getAmount());
+            assertEquals("m", buildConfig.getSpec().getResources().getLimits().get("cpu").getFormat());
         });
     }